LPS-139 Avoid PortalException when deleting the CommerceChannelGroup

+++ modules/apps/commerce/commerce-service/src/main/java/com/liferay/commerce/internal/model/listener/CommerceChannelModelListener.java
@@ -34,6 +34,10 @@ public class CommerceChannelModelListener
 
 	@Override
 	public void onBeforeRemove(CommerceChannel commerceChannel) {
+		if (commerceChannel.getGroupId() == 0) {
+			return;
+		}
+
 		try {
 			_commerceShippingMethodLocalService.deleteCommerceShippingMethods(
 				commerceChannel.getGroupId());
